calls-manager: Connect listener past the list model filters
The list model filters connect signal handlers for item-changed to
refilter the list models.

In order to enable protocols based on what the origin supports
`on_origins_changed` needs these filtered results. Since it is a signal
handler by itself connect it late so it runs later in the filter chain.

Closes: https://gitlab.gnome.org/GNOME/calls/-/issues/601