ci: Rename dist tarball so it will be found by the release pipeline
The TARBALL_ARTIFACT_PATH is formed from the name of the tag,
which is prefixed by a "v". Additionally special characters like the
tilde (e.g. 48~beta.1) are translated to underscores.

This resulted in the TARBALL_ARTIFACT_PATH Gitlab variable
not actually pointing to the generated tarball and therefore
a failing release job.

While I'd have preferred massaging the gitlab variable into the right
form, doing string substitution is not working on gitlab variable,
so I opted for a this workaround instead.

Fixes: aa8c1b571d4901135539092a247c91d955a89e8a

Signed-off-by: Evangelos Ribeiro Tzaras <devrtz@fortysixandtwo.eu>
Part-of: <https://gitlab.gnome.org/GNOME/calls/-/merge_requests/765>