new-call-box: wrap dialer in scrolled window
It was possible to increase the window's size below the
minimum size of the dialer view. As such it is important
to wrap it in a scrolled window in order not to lose
access to otherwise hidden interface components.

This is, of course, not a replacement for a better
adapting dialer.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>
Part-of: <https://gitlab.gnome.org/GNOME/calls/-/merge_requests/791>